Should have been in commit 24393c91
We changed the uint16_t to uint32_t to get around the warning comparing those variables as greater than 65535, but we never changed the “%u” in the format strings to “%lu”.

This just casts those uint32_t back to uint16_t.

+++ src/DFR_Radar.cpp
@@ -260,7 +260,7 @@ bool DFR_Radar::setOutputLatency( float triggerDelay, float resetDelay )
     return false;
 
   char _comOutputLatency[29] = {0};
-  sprintf( _comOutputLatency, comOutputLatency, _triggerDelay , _resetDelay );
+  sprintf( _comOutputLatency, comOutputLatency, (uint16_t)_triggerDelay , (uint16_t)_resetDelay );
 
   return setConfig( _comOutputLatency );
 }
